todo_api v2
In this approach we only generate the basic structure. This is better than the generic approach `mix phx.new todo_api`, when we only want to create a `REST API` using Phoenix and don’t need LiveView, assets, HTML, dashboard, or mailer. It helps you avoid generating unnecessary files and dependencies that are not required for an API-only application.